+++ packages/features/ee/workflows/pages/workflow.tsx
@@ -1,10 +1,8 @@
 import { zodResolver } from "@hookform/resolvers/zod";
 import type { WorkflowStep } from "@prisma/client";
-import { isValidPhoneNumber } from "libphonenumber-js";
 import { useSession } from "next-auth/react";
 import { useEffect, useState } from "react";
 import { useForm } from "react-hook-form";
-import { z } from "zod";
 
 import Shell, { ShellMain } from "@calcom/features/shell/Shell";
 import { classNames } from "@calcom/lib";
@@ -12,14 +10,8 @@ import { SENDER_ID } from "@calcom/lib/constants";
 import { useLocale } from "@calcom/lib/hooks/useLocale";
 import { useParamsWithFallback } from "@calcom/lib/hooks/useParamsWithFallback";
 import { HttpError } from "@calcom/lib/http-error";
-import {
-  MembershipRole,
-  TimeUnit,
-  WorkflowActions,
-  WorkflowTemplates,
-  WorkflowTriggerEvents,
-} from "@calcom/prisma/enums";
-import { stringOrNumber } from "@calcom/prisma/zod-utils";
+import type { TimeUnit, WorkflowTriggerEvents } from "@calcom/prisma/enums";
+import { MembershipRole, WorkflowActions } from "@calcom/prisma/enums";
 import type { RouterOutputs } from "@calcom/trpc/react";
 import { trpc } from "@calcom/trpc/react";
 import useMeQuery from "@calcom/trpc/react/hooks/useMeQuery";
@@ -30,6 +22,7 @@ import LicenseRequired from "../../common/components/LicenseRequired";
 import SkeletonLoader from "../components/SkeletonLoaderEdit";
 import WorkflowDetailsPage from "../components/WorkflowDetailsPage";
 import { isSMSAction, isSMSOrWhatsappAction } from "../lib/actionHelperFunctions";
+import { formSchema, querySchema } from "../lib/schema";
 import { getTranslatedText, translateVariablesToEnglish } from "../lib/variableTranslations";
 
 export type FormValues = {
@@ -41,50 +34,6 @@ export type FormValues = {
   timeUnit?: TimeUnit;
   selectAll: boolean;
 };
-
-export function onlyLettersNumbersSpaces(str: string) {
-  if (str.length <= 11 && /^[A-Za-z0-9\s]*$/.test(str)) {
-    return true;
-  }
-  return false;
-}
-
-const formSchema = z.object({
-  name: z.string(),
-  activeOn: z.object({ value: z.string(), label: z.string() }).array(),
-  trigger: z.nativeEnum(WorkflowTriggerEvents),
-  time: z.number().gte(0).optional(),
-  timeUnit: z.nativeEnum(TimeUnit).optional(),
-  steps: z
-    .object({
-      id: z.number(),
-      stepNumber: z.number(),
-      action: z.nativeEnum(WorkflowActions),
-      workflowId: z.number(),
-      reminderBody: z.string().nullable(),
-      emailSubject: z.string().nullable(),
-      template: z.nativeEnum(WorkflowTemplates),
-      numberRequired: z.boolean().nullable(),
-      includeCalendarEvent: z.boolean().nullable(),
-      sendTo: z
-        .string()
-        .refine((val) => isValidPhoneNumber(val) || val.includes("@"))
-        .optional()
-        .nullable(),
-      sender: z
-        .string()
-        .refine((val) => onlyLettersNumbersSpaces(val))
-        .optional()
-        .nullable(),
-      senderName: z.string().optional().nullable(),
-    })
-    .array(),
-  selectAll: z.boolean(),
-});
-
-const querySchema = z.object({
-  workflow: stringOrNumber,
-});
 
 function WorkflowPage() {
   const { t, i18n } = useLocale();
@@ -129,10 +78,6 @@ function WorkflowPage() {
     teamId: workflow?.team?.id,
   });
 
-  // const { data: eventTypeGroups, isPending: isPendingEventTypes } =
-  //   trpc.viewer.eventTypes.getByViewer.useQuery();
-
-  const { data: otherTeams, isPending: isPendingTeams } = trpc.viewer.organizations.listOtherTeams.useQuery();
   const isOrg = workflow?.team?.isOrganization ?? false;
 
   const teamId = workflow?.teamId ?? undefined;
@@ -144,62 +89,13 @@ function WorkflowPage() {
   console.log("trpc.viewer.eventTypes.getEventTypeOptions", data);
   console.log("selectedOptions", selectedOptions);
 
-  // const profileTeamsOptions =
-  //   isOrg && eventTypeGroups
-  //     ? eventTypeGroups?.profiles
-  //         .filter((profile) => !!profile.teamId)
-  //         .map((profile) => {
-  //           return {
-  //             value: String(profile.teamId) || "",
-  //             label: profile.name || profile.slug || "",
-  //           };
-  //         })
-  //     : [];
+  const teamOptions = data?.teamOptions ?? [];
 
-  const profileTeamsOptions = isOrg && data?.profilesTeamsOptions ? profilesTeamsOptions : [];
-
-  const otherTeamsOptions = otherTeams
-    ? otherTeams.map((team) => {
-        return {
-          value: String(team.id) || "",
-          label: team.name || team.slug || "",
-        };
-      })
-    : [];
-
-  const teamOptions = profileTeamsOptions.concat(otherTeamsOptions);
-
-  // const eventTypeOptions = useMemo(
-  //   () =>
-  //     eventTypeGroups?.eventTypeGroups.reduce((options, group) => {
-  //       /** don't show team event types for user workflow */
-  //       if (!teamId && group.teamId) return options;
-  //       /** only show correct team event types for team workflows */
-  //       if (teamId && teamId !== group.teamId) return options;
-  //       return [
-  //         ...options,
-  //         ...group.eventTypes
-  //           .filter(
-  //             (evType) =>
-  //               !evType.metadata?.managedEventConfig ||
-  //               !!evType.metadata?.managedEventConfig.unlockedFields?.workflows ||
-  //               !!teamId
-  //           )
-  //           .map((eventType) => ({
-  //             value: String(eventType.id),
-  //             label: `${eventType.title} ${
-  //               eventType.children && eventType.children.length ? `(+${eventType.children.length})` : ``
-  //             }`,
-  //           })),
-  //       ];
-  //     }, [] as Option[]) || [],
-  //   [eventTypeGroups]
-  // );
   let allEventTypeOptions = data?.eventTypeOptions ?? [];
   const distinctEventTypes = new Set();
 
   if (!teamId && isMixedEventType) {
-    allEventTypeOptions = [...data?.eventTypeOptions, ...selectedOptions];
+    allEventTypeOptions = [...allEventTypeOptions, ...selectedOptions];
     allEventTypeOptions = allEventTypeOptions.filter((option) => {
       const duplicate = distinctEventTypes.has(option.value);
       distinctEventTypes.add(option.value);
@@ -207,12 +103,11 @@ function WorkflowPage() {
     });
   }
 
-  // console.log("allEventTypeOptions", allEventTypeOptions);
   const readOnly =
     workflow?.team?.members?.find((member) => member.userId === session.data?.user.id)?.role ===
     MembershipRole.MEMBER;
 
-  const isPending = isPendingWorkflow || isPendingEventTypes || isPendingTeams;
+  const isPending = isPendingWorkflow || isPendingEventTypes;
 
   useEffect(() => {
     if (!isPending) {
