+++ packages/lib/server/repository/membership.ts
@@ -165,6 +165,41 @@ export class MembershipRepository {
       })
     );
 
+    // const select = Prisma.validator<Prisma.MembershipSelect>()({
+    //   id: true,
+    //   teamId: true,
+    //   userId: true,
+    //   accepted: true,
+    //   role: true,
+    //   disableImpersonation: true,
+    //   team: {
+    //     include: {
+    //       parent: {
+    //         select: teamParentSelect,
+    //       },
+    //       ...(!skipEventTypes
+    //         ? {
+    //             eventTypes: {
+    //               select: {
+    //                 ...eventTypeSelect,
+    //                 hashedLink: true,
+    //                 children: { select: { id: true } },
+    //               },
+    //               orderBy: [
+    //                 {
+    //                   position: "desc",
+    //                 },
+    //                 {
+    //                   id: "asc",
+    //                 },
+    //               ],
+    //             },
+    //           }
+    //         : {}),
+    //     },
+    //   },
+    // });
+
     return await prisma.membership.findMany({
       where: prismaWhere,
       include: {
@@ -179,7 +214,7 @@ export class MembershipRepository {
                     select: {
                       ...eventTypeSelect,
                       hashedLink: true,
-                      children: true,
+                      children: { select: { id: true } },
                     },
                     orderBy: [
                       {
